test: break out WsPromiser into a separate file
