+++ test/history.js
@@ -3,9 +3,9 @@ chai.Should()
 chai.use(require('chai-things'))
 const _ = require('lodash')
 const freeport = require('freeport-promise')
-const WebSocket = require('ws')
 const Server = require('../lib')
 const fetch = require('node-fetch')
+const { WsPromiser } = require('./servertestutilities')
 
 let testDelta = {
   updates: [
@@ -25,7 +25,7 @@ let testDelta = {
 
 let dummyHistoryProvider = app => {
   return {
-    streamHistory: (cookie, query, onDelta) => {
+    streamHistory: (cookie, options, onDelta) => {
       setTimeout(() => {
         testDelta.context = `vessels.${app.selfId}`
         onDelta(testDelta)
@@ -36,6 +36,9 @@ let dummyHistoryProvider = app => {
     getHistory: (date, path, cb) => {
       testDelta.context = `vessels.${app.selfId}`
       cb([testDelta])
+    },
+    hasAnyData: (options, cb) => {
+      cb(true)
     }
   }
 }
@@ -87,37 +90,3 @@ describe('History', _ => {
     json.should.have.nested.property('performance.velocityMadeGood')
   })
 })
-
-// Connects to the url via ws
-// and provides Promises that are either resolved within
-// timeout period as the next message from the ws or
-// the string "timeout" in case timeout fires
-function WsPromiser (url) {
-  this.ws = new WebSocket(url)
-  this.ws.on('message', this.onMessage.bind(this))
-  this.callees = []
-}
-
-WsPromiser.prototype.nextMsg = function () {
-  const callees = this.callees
-  return new Promise((resolve, reject) => {
-    callees.push(resolve)
-    setTimeout(_ => {
-      resolve('timeout')
-    }, 250)
-  })
-}
-
-WsPromiser.prototype.onMessage = function (message) {
-  const theCallees = this.callees
-  this.callees = []
-  theCallees.forEach(callee => callee(message))
-}
-
-WsPromiser.prototype.send = function (message) {
-  const that = this
-  return new Promise((resolve, reject) => {
-    that.ws.send(JSON.stringify(message))
-    setTimeout(() => resolve('wait over'), 100)
-  })
-}
