+++ test/subscriptions.js
@@ -3,7 +3,7 @@ const assert = require('assert')
 const freeport = require('freeport-promise')
 const WebSocket = require('ws')
 const rp = require('request-promise')
-const startServerP = require('./servertestutilities').startServerP
+const { startServerP, WsPromiser } = require('./servertestutilities')
 
 function getDelta (overwrite) {
   const delta = {
@@ -482,37 +482,3 @@ describe('Subscriptions', _ => {
       })
   })
 })
-
-// Connects to the url via ws
-// and provides Promises that are either resolved within
-// timeout period as the next message from the ws or
-// the string "timeout" in case timeout fires
-function WsPromiser (url) {
-  this.ws = new WebSocket(url)
-  this.ws.on('message', this.onMessage.bind(this))
-  this.callees = []
-}
-
-WsPromiser.prototype.nextMsg = function () {
-  const callees = this.callees
-  return new Promise((resolve, reject) => {
-    callees.push(resolve)
-    setTimeout(_ => {
-      resolve('timeout')
-    }, 250)
-  })
-}
-
-WsPromiser.prototype.onMessage = function (message) {
-  const theCallees = this.callees
-  this.callees = []
-  theCallees.forEach(callee => callee(message))
-}
-
-WsPromiser.prototype.send = function (message) {
-  const that = this
-  return new Promise((resolve, reject) => {
-    that.ws.send(JSON.stringify(message))
-    setTimeout(() => resolve('wait over'), 100)
-  })
-}
